resync; also added some Python stuff that's not in SFS
git-svn-id: svn://okws.org/trunk/sfslite2@912 7d287422-6eea-da11-b677-00123f2a783a

+++ async/itree.h
@@ -87,7 +87,7 @@ protected:
   itree_core (const itree_core &);
   itree_core &operator = (const itree_core &);
 
-#define eos ((int) &(((T *) 0)->*field).p)
+#define eos ((ptrdiff_t) &(((T *) 0)->*field).p)
 #define cmpfn scmp, (void *) this
 
 public:
